requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
python-dotenv==1.0.0
